Update actions to use dynamic or predefined inputs
